wip: bare popup visible
