+++ app/src/ctx.rs
@@ -1,3 +1,4 @@
+use core::popup::Popup;
 use core::question::Questions;
 use core::topic::Topic;
 use leetcode_db::Db;
@@ -5,14 +6,15 @@ use leetcode_db::Db;
 pub struct Ctx {
     pub topic: Topic,
     pub question: Questions,
+    pub popup: Popup,
 }
 
 impl Ctx {
     pub(super) async fn new(db: &Db) -> Self {
-        let t = Self {
+        Self {
             topic: Topic::new(db).await,
             question: Questions::new(),
-        };
-        t
+            popup: Default::default(),
+        }
     }
 }
